refactor(hooks): structure dir structure
